benchmark/DeepTIO/requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
cloudpickle==1.6.0
cryptography==2.1.4
cycler==0.10.0
decorator==5.0.9
gast==0.2.2
grpcio==1.22.0
h5py==2.9.0
idna==2.6
Keras==2.2.4
Keras-Applications==1.0.8
keras-mdn-layer==0.3.0
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
Markdown==3.1.1
matplotlib==3.3.4
mock==3.0.5
numpy==1.16.4
opencv-python==4.4.0.42
Pillow==8.3.1
protobuf==3.9.0
pycrypto==2.6.1
pygobject==3.26.1
pyparsing==2.4.7
python-apt==1.6.4
python-dateutil==2.8.2
pyxdg==0.25
PyYAML==5.4.1
scipy==1.5.4
SecretStorage==2.3.1
six==1.11.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.2
tensorflow-probability==0.6.0
termcolor==1.1.0
Werkzeug==0.15.4
